fixed int issues when it should be float

+++ server/server.py
@@ -209,7 +209,7 @@ try:
         #get entry from http body
         message = request.forms.get('entry')
         input_vector_clock = json.loads(request.forms.get('vector_clock'))
-        time_stamp = int(request.forms.get('time_stamp'))
+        time_stamp = float(request.forms.get('time_stamp'))
         print "the action is", action
         update_vector_clock(input_vector_clock)
         new_input = Element(action, int(element_id), message, input_vector_clock, time_stamp)
